Fix the movement service.... again
This PR fixes the movement service based on many hours of painful
deliberation on the correct thing to do. It now:

 * Tries to fix allocations with a nil prison
 * Handle REL correctly
 * Handles ADM correctly for transfers
 * Ignores newly arriving offenders

+++ app/services/movement_service.rb
@@ -19,30 +19,43 @@ class MovementService
     movements
   end
 
-  # rubocop:disable Metrics/CyclomaticComplexity
   def self.process_movement(movement)
     if movement.movement_type == Nomis::Models::MovementType::RELEASE
       return process_release(movement)
     end
 
-    # We need to check whether the from_agency is from without the prison estate
+    # We need to check whether the from_agency is from within the prison estate
     # to know whether it is a transfer.  If it isn't then we want to bail and
     # not process the new admission.
-    if movement.movement_type == Nomis::Models::MovementType::ADMISSION &&
-      movement.direction_code == Nomis::Models::MovementDirection::IN &&
-      movement.from_agency.present? &&
-      a_transfer?(movement.from_agency) &&
-      movement.to_agency.present?
+    if [Nomis::Models::MovementType::ADMISSION,
+        Nomis::Models::MovementType::TRANSFER].include?(movement.movement_type)
+
       return process_transfer(movement)
     end
 
     false
   end
-# rubocop:enable Metrics/CyclomaticComplexity
 
 private
 
   def self.process_transfer(transfer)
+    # Bail if this is a new admission to prison
+    return false unless transfer.from_prison? && transfer.to_prison?
+    return false unless transfer.direction_code == 'IN'
+
+    # We only want to deallocate the offender if they have not already been
+    # allocated at their new prison
+    if AllocationVersion.where(
+      nomis_offender_id: transfer.offender_no,
+      prison: transfer.to_agency
+    ).count > 0
+
+      Rails.logger.info("Offender #{transfer.offender_no} was transferred but \
+        an allocation at #{transfer.to_agency} already exists")
+
+      return false
+    end
+
     Rails.logger.info("Processing transfer for #{transfer.offender_no}")
 
     return false unless should_process?(transfer.offender_no)
@@ -56,16 +69,14 @@ private
   # case information (in case they come back one day), and we
   # should de-activate any current allocations.
   def self.process_release(release)
+    return false unless release.to_agency == 'OUT' && release.from_prison?
+
     Rails.logger.info("Processing release for #{release.offender_no}")
     CaseInformationService.delete_information(release.offender_no)
     AllocationVersion.deallocate_offender(release.offender_no,
                                           AllocationVersion::OFFENDER_RELEASED)
 
     true
-  end
-
-  def self.a_transfer?(from_agency_code)
-    PrisonService::PRISONS.include?(from_agency_code)
   end
 
   def self.should_process?(offender_no)
